requirements.txt
tinydb
argparse
beautifulsoup4
requests
pillow
pandas
